fix(ou-tree): move .js->.ts jest resolver mapping to shared jest configs
The package-local jest.config.js only fixed `yarn workspace @dhis2-ui/ou-tree test`;
the root `yarn test` (used by CI) still failed to resolve `.js` sibling imports from
`.ts`/`.tsx` files. Add the moduleNameMapper (re-spreading cli-app-scripts' default
mocks) to jest.config.shared.js and jest.config.js instead, and restore ou-tree's
conventional test script now that its own jest.config.js is no longer needed.
