+++ components/ou-tree/package.json
@@ -22,7 +22,7 @@
     "scripts": {
         "start": "storybook dev -c ../../storybook/config --port 5000",
         "build": "d2-app-scripts build && tsc --project tsconfig.build.json",
-        "test": "d2-app-scripts test --jestConfig ./jest.config.js",
+        "test": "d2-app-scripts test --jestConfig ../../jest.config.shared.js",
         "typecheck": "tsc --noEmit"
     },
     "peerDependencies": {
